requirements.txt
requests==2.32.3
python-dotenv==1.0.1
jinja2==3.1.4
pytz==2024.2
boto3==1.34.128
hvac==2.3.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
msal==1.29.0
